Implementing Data rows in cloud tests to avoid configuration file complexity. Moving cloud configuration provider files.
