+++ LibsNoSamples.sln
@@ -49,8 +49,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Identity.Client",
 		{528C5488-C0D8-45C3-BEAC-3E52C22631C9} = {528C5488-C0D8-45C3-BEAC-3E52C22631C9}
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Identity.Test.ConfigurationProvider", "tests\Microsoft.Identity.Test.ConfigurationProvider\Microsoft.Identity.Test.ConfigurationProvider.csproj", "{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -112,10 +110,6 @@ Global
 		{60117A9B-4BB8-472E-BFFF-52CBF67CA95A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{60117A9B-4BB8-472E-BFFF-52CBF67CA95A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{60117A9B-4BB8-472E-BFFF-52CBF67CA95A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -135,7 +129,6 @@ Global
 		{CFC2232F-3276-42A4-B8E3-F03B6F8BAA82} = {7930A443-1981-49CD-8AF5-02E25F3A7FF3}
 		{E7F9B699-4941-4053-AB81-7E26C4D24805} = {5D231979-0A13-456D-B9F9-CED41F935FD9}
 		{3DC6EC76-D350-4D43-B206-A4CEC8AA36D4} = {5D231979-0A13-456D-B9F9-CED41F935FD9}
-		{22DEC4BF-29B5-40B3-AB24-8E8B19FA5C75} = {5D231979-0A13-456D-B9F9-CED41F935FD9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6C88EF77-AF03-4F82-B666-1831DABAA119}
