--- tests/Microsoft.Identity.Test.ConfigurationProvider/CloudType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace Microsoft.Identity.Test.ConfigurationProvider
-{
-    public enum CloudType
-    {
-        AzureCloud,
-        AzureGermanyCloud,
-        AzureUSGovernment,
-        AzureChinaCloud,
-        AzureUSGovernmentArlington
-    }
-}
