--- tests/Microsoft.Identity.Test.Integration/CloudSelection.json
@@ -1,3 +1,0 @@
-{
-  "CloudType": "DefaultCloud"
-}
