+++ tests/Microsoft.Identity.Test.Integration/HeadlessTests/UsernamePasswordIntegrationTests.cs
@@ -24,12 +24,14 @@
 // THE SOFTWARE.
 //
 //------------------------------------------------------------------------------
+
 #if !WINDOWS_APP && !ANDROID && !iOS // U/P not available on UWP, Android and iOS
 using System.Net;
 using System.Security;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Test.Common;
+using Microsoft.Identity.Test.Integration.Infrastructure;
 using Microsoft.Identity.Test.LabInfrastructure;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -57,82 +59,53 @@ namespace Microsoft.Identity.Test.Integration.HeadlessTests
 
         #region Happy Path Tests
         [TestMethod]
-        [TestCategory("AzureUSGovernment")]
         public async Task ROPC_AAD_Async()
         {
             var labResponse = LabUserHelper.GetDefaultUser();
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         [TestMethod]
         public async Task ROPC_ADFSv4Federated_Async()
         {
             var labResponse = LabUserHelper.GetAdfsUser(FederationProvider.AdfsV4, true);
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         [TestMethod]
         public async Task ROPC_ADFSv4Managed_Async()
         {
             var labResponse = LabUserHelper.GetAdfsUser(FederationProvider.AdfsV4, false);
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         [TestMethod]
-        [TestCategory("AzureUSGovernment")]
         public async Task ROPC_ADFSv3Federated_Async()
         {
             var labResponse = LabUserHelper.GetAdfsUser(FederationProvider.AdfsV3, true);
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         [TestMethod]
-        [TestCategory("AzureUSGovernment")]
         public async Task ROPC_ADFSv3Managed_Async()
         {
             var labResponse = LabUserHelper.GetAdfsUser(FederationProvider.AdfsV3, false);
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         [TestMethod]
         public async Task ROPC_ADFSv2Fderated_Async()
         {
             var labResponse = LabUserHelper.GetAdfsUser(FederationProvider.AdfsV2, true);
-            await RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
+            await SeleniumTestHelper.RunHappyPathTestAsync(labResponse).ConfigureAwait(false);
         }
 
         #endregion
 
         [TestMethod]
-        [TestCategory("AzureUSGovernment")]
         public async Task AcquireTokenWithManagedUsernameIncorrectPasswordAsync()
         {
-            var labResponse = LabUserHelper.GetDefaultUser();
-            var user = labResponse.User;
-
-            SecureString incorrectSecurePassword = new SecureString();
-            incorrectSecurePassword.AppendChar('x');
-            incorrectSecurePassword.MakeReadOnly();
-
-            PublicClientApplication msalPublicClient = new PublicClientApplication(labResponse.AppId, _authority);
-
-            try
-            {
-                var result =
-                     await msalPublicClient.AcquireTokenByUsernamePasswordAsync(_scopes, user.Upn, incorrectSecurePassword)
-                     .ConfigureAwait(false);
-            }
-            catch (MsalServiceException ex)
-            {
-                Assert.IsTrue(!string.IsNullOrWhiteSpace(ex.CorrelationId));
-                Assert.AreEqual(400, ex.StatusCode);
-                Assert.AreEqual("invalid_grant", ex.ErrorCode);
-                Assert.IsTrue(ex.Message.StartsWith("AADSTS50126: Invalid username or password"));
-
-                return;
-            }
-
-            Assert.Fail("Bad exception or no exception thrown");
+            await SeleniumTestHelper.AcquireTokenWithManagedUsernameIncorrectPasswordTestAsync().ConfigureAwait(false);
         }
 
         [TestMethod]
@@ -157,30 +130,6 @@ namespace Microsoft.Identity.Test.Integration.HeadlessTests
 
             var result = Assert.ThrowsExceptionAsync<MsalException>(async () =>
                  await msalPublicClient.AcquireTokenByUsernamePasswordAsync(_scopes, user.Upn, incorrectSecurePassword).ConfigureAwait(false));
-        }
-
-        private async Task RunHappyPathTestAsync(LabResponse labResponse)
-        {
-            var user = labResponse.User;
-
-            SecureString securePassword = new NetworkCredential("", user.Password).SecurePassword;
-
-            PublicClientApplication msalPublicClient = new PublicClientApplication(labResponse.AppId, _authority);
-
-            //AuthenticationResult authResult = await msalPublicClient.AcquireTokenByUsernamePasswordAsync(Scopes, user.Upn, securePassword).ConfigureAwait(false);
-            AuthenticationResult authResult = await msalPublicClient.AcquireTokenByUsernamePasswordAsync(_scopes, user.Upn, securePassword).ConfigureAwait(false);
-
-            Assert.IsNotNull(authResult);
-            Assert.IsNotNull(authResult.AccessToken);
-            Assert.IsNotNull(authResult.IdToken);
-            Assert.AreEqual(user.Upn, authResult.Account.Username);
-            // If test fails with "user needs to consent to the application, do an interactive request" error,
-            // Do the following: 
-            // 1) Add in code to pull the user's password before creating the SecureString, and put a breakpoint there.
-            // string password = ((LabUser)user).GetPassword();
-            // 2) Using the MSAL Desktop app, make sure the ClientId matches the one used in integration testing.
-            // 3) Do the interactive sign-in with the MSAL Desktop app with the username and password from step 1.
-            // 4) After successful log-in, remove the password line you added in with step 1, and run the integration test again.
         }
     }
 }
