+++ tests/Microsoft.Identity.Test.LabInfrastructure/UserQueryParameters.cs
@@ -27,7 +27,7 @@
 
 using System;
 using System.Collections.Generic;
-using Microsoft.Identity.Test.ConfigurationProvider;
+using Microsoft.Identity.Test.LabInfrastructure.CloudInfrastructure;
 
 namespace Microsoft.Identity.Test.LabInfrastructure
 {
@@ -41,11 +41,13 @@ namespace Microsoft.Identity.Test.LabInfrastructure
         public UserType? UserType { get; set; }
         public bool? IsExternalUser { get; set; }
         public B2CIdentityProvider? B2CIdentityProvider { get; set; }
-        public CloudType CloudType { get; set; }
 
-        public UserQuery()
+        public CloudType CloudType
         {
-            CloudType = CloudConfigurationProvider.CloudType;
+            get
+            {
+                return CloudConfigurationProvider.CloudType;
+            }
         }
 
         public static UserQuery DefaultUserQuery => new UserQuery
